requirements.txt
scikit-learn
pandas
numpy
matplotlib
seaborn
streamlit
joblib
Pillow
